requirements.txt
maya==0.1.4
pystache==0.5.4
PyYAML==3.12
zipa==0.3.0
